Guard against missing original file during install
During OTA update, when installing files to /boot/, backup is attempted
regardless of whether the target file already exists or not. If the target file
does not exist already, this results in failure, even though continuing the
install is a legitimate action. This patch fixes the installer such that
missing target paths do not result in failure.

With this patch, the installer gains the ability to add files to /boot/ that
previously did not exist (e.g., if we want to add a separate initramfs, and so
on).

This patch also adds a similar test when restoring backups, so that restore is
only attempted in cases where backup actually exists.

+++ rxos/local/ota-pkg/src/installer.in.sh
@@ -44,12 +44,16 @@ extract_to_bootfs() {
   filename="$1"
   $LOG "Installing $filename"
   mount -o remount,rw /boot || fail "Could not unlock boot partition"
-  cp "/boot/$filename" "/boot/${filename}.backup" \
-    || fail "Could not back up /boot/${filename}"
-  sync
-  if ! $INSTALLER --extract "$filename" /boot; then
-    mv "/boot/${filename}.backup" "/boot/$filename"
+  if [ -f "/boot/$filename" ]; then
+    cp "/boot/$filename" "/boot/${filename}.backup" \
+      || fail "Could not back up /boot/${filename}"
     sync
+  fi
+  if ! $INSTALLER --extract "$filename" /boot; then
+    if [ -f "/boot/${filename}.backup" ]; then
+      mv "/boot/${filename}.backup" "/boot/$filename"
+      sync
+    fi
     fail "Could not extract $filename"
   fi
   sync
